pyproject.toml
[tool.poetry]
name = "pythondigest"
version = "0.1.0"
description = ""
authors = ["axsapronov <a@sapronov.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = {extras = ["argon2"], version = "^4.1.3"}
django-bootstrap-form = ">=3.4"
django-digg-paginator = ">=0.2.0"
django-user-accounts = ">=2.1.0"
django-taggit-autosuggest = ">=0.4.1"
django-taggit = ">=0.24.0"
django-modeladmin-reorder = ">=0.3.1"
django-ckeditor = ">=5.9.0"
django-siteblocks = ">=0.6.1"
django-cachalot = ">=2.1.0"
django-compressor = ">=4.1"
django-secretballot = ">=2.0.0"
django-likes = {git = "https://github.com/axsapronov/django-likes.git", rev = "django4-signals"}
django-debug-toolbar = ">=1.11.1"
django-remdow = ">=0.0.7"
requests = ">=2.21.0"
lxml = ">=4.3.5"
pyyaml = ">=5.2"
beautifulsoup4 = ">=4.10.0"
feedparser = ">=5.2.1"
pytils = ">=0.3"
sorl-thumbnail = ">=12.7.0"
stem = ">=1.8.1"
urlobject = ">=2.4.3"
micawber = ">=0.5.4"
rsa = ">=4.4"
twx-botapi = "3.6.1"
google-api-python-client = ">=1.12.11"
tweepy = ">=3.7.0"
funcy = ">=1.17"
pillow = ">=5.4.1"
unidecode = ">=1.3.6"
readability-lxml = {git = "https://github.com/buriy/python-readability.git", rev = "65f2f13"}
psycopg2-binary = "^2.9.5"
pymemcache = "^4.0.0"
django-browser-reload = "^1.6.0"
django-bootstrap3 = "^22.2"
django-yaturbo = "^1.0.1"
bleach = "^5.0.1"
django-environ = "^0.9.0"
django-redis = "^5.2.0"
sentry-sdk = "^1.11.1"
django-letsencrypt = "^4.1.0"
django-cache-memoize = "^0.1.10"
django-htmlmin = "^0.11.0"
uwsgi = "^2.0.21"
whitenoise = {extras = ["brotli"], version = "^6.2.0"}

[tool.poetry.group.test.dependencies]
django-stubs = "^1.13.0"
pytest = "^7.1.3"
pytest-sugar = "^0.9.6"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"
mock = ">=3.0.5"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
flake8-isort = "^5.0.0"
coverage = "^6.5.0"
black = "^22.10.0"
pylint-django = "^2.5.3"
pre-commit = "^2.20.0"
django-extra-checks = "^0.13.0"
django-migration-linter = "^4.1.0"
autoflake = "^1.7.7"

[tool.black]
line-length = 119
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 119
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
ensure_newline_before_comments = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
